+++ mne/io/__init__.py
@@ -33,5 +33,4 @@ from .fiff import read_raw_fif
 from .fiff import RawFIF
 from .fiff import RawFIF as Raw
 from .base import concatenate_raws
-from .chpi import  get_chpi_positions
 from .reference import set_eeg_reference, set_bipolar_reference, add_reference_channels
